# api/docs/swagger.yaml
definitions:
...
  genproto.PersonalEventsCreateReq:
    properties:
      date:
        type: string
      preview:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.PersonalEventsGetAllRes:
    properties:
      count:
        type: integer
      events:
        items:
          $ref: '#/definitions/genproto.PersonalEventsRes'
        type: array
    type: object
  genproto.PersonalEventsGetByIdRes:
    properties:
      event:
        $ref: '#/definitions/genproto.PersonalEventsRes'
    type: object
  genproto.PersonalEventsRes:
    properties:
      date:
        type: string
      id:
        type: string
      preview:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.SharedMemoriesCreateReq:
    properties:
      memory_id:
        type: string
      message:
        type: string
      recipient_id:
        type: string
      shared_id:
        type: string
    type: object
  genproto.SharedMemoriesGetAllRes:
    properties:
      count:
        type: integer
      memories:
        items:
          $ref: '#/definitions/genproto.SharedMemoriesRes'
        type: array
    type: object
  genproto.SharedMemoriesGetByIdRes:
    properties:
      memory:
        $ref: '#/definitions/genproto.SharedMemoriesRes'
    type: object
  genproto.SharedMemoriesRes:
    properties:
      id:
        type: string
      memory_id:
        type: string
      message:
        type: string
      recipient_id:
        type: string
      shared_at:
        type: string
      shared_id:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /memory/{id}/shared:
    post:
      consumes:
      - application/json
      description: Shared a new memory
      parameters:
      - description: SharedMemory data
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/genproto.SharedMemoriesCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: SharedMemory created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Shared memory
      tags:
      - shared
  /memory/{id}/shared/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a memory by ID
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory deleted
          schema:
            type: string
        "400":
          description: Invalid memory ID
          schema:
            type: string
        "404":
          description: Memory not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete memory
      tags:
      - shared
    get:
      consumes:
      - application/json
      description: Get a memory
      parameters:
      - description: SharedMemory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.SharedMemoriesGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get memory
      tags:
      - shared
    put:
      consumes:
      - application/json
      description: Update an existing shared memory
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Message
        in: query
        name: message
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Memory not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update shared memory
      tags:
      - shared
  /memory/{id}/shared/all:
    get:
      consumes:
      - application/json
      description: Get all memories
      parameters:
      - description: SharedId
        in: query
        name: shared_id
        type: string
      - description: RecipientId
        in: query
        name: recipient_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.SharedMemoriesGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all memories
      tags:
      - shared
...
  /timeline/personal:
    post:
      consumes:
      - application/json
      description: Create a new event
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/genproto.PersonalEventsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Event created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - personal
  /timeline/personal/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted
          schema:
            type: string
        "400":
          description: Invalid event ID
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - personal
    get:
      consumes:
      - application/json
      description: Get a event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PersonalEventsGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - personal
    put:
      consumes:
      - application/json
      description: Update an existing event
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Preview
        in: query
        name: preview
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - personal
  /timeline/personal/all:
    get:
      consumes:
      - application/json
      description: Get all event
      parameters:
      - description: Date
        in: query
        name: date
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PersonalEventsGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all event
      tags:
      - personal
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"